It is now the torso which is the main sprite of the character.
Exactly why did I do this? Well first of all because it makes sense. But also because it makes the larger of the sprites into the object which physics is calculated on. Thirdly because it makse sense that the hookshot eminates from the torso and not the feet.

In other news the hookshot now eminates from outside of of bobby, not his guts.