Added a demo to the start menu.
Sadly it is not stable. I have no idea why but it runs differently about every time it's run. It could have to do with the faulting timing of Phasers timing engine. Or that physics updates are run with different deltas every time the game is run.

This variable update delta is becoming a real issue for us. I would much rather have it so that the update delta is fixed and that frames are being dropped in favour of reliable physics calculations.